fix: refactor emit notify events into service function
