+++ lib/components/services/notify/index.js
@@ -1,12 +1,15 @@
 const schema = require('./schema.json')
 const MessageTemplate = require('./Message-template')
 
+const moment = require('moment')
+
 class NotifyService {
   boot (options) {
     this.notificationModel = options.bootedServices.storage.models.tymly_govUkNotifications
     this.customTemplateModel = options.bootedServices.storage.models.tymly_govUkCustomTemplates
+    this.sendConfigModel = options.bootedServices.storage.models.tymly_customTemplateSendConfig
     this.recipientModel = options.bootedServices.storage.models.tymly_govUkCustomTemplateRecipients
-
+    this.scheduleService = options.bootedServices.schedule
     this.notifyClient = {}
     this.templates = {}
     this.events = {}
@@ -49,6 +52,121 @@ class NotifyService {
       }
     }
   } // boot
+
+  async emitNotifyEvents (eventNames = [], parameters = {}, recipientsFromStateMachine = []) {
+    if (!Array.isArray(eventNames)) {
+      console.log('EmitNotifyEvents: eventNames must be an array of strings')
+      return
+    }
+
+    const sent = []
+
+    for (const eventName of eventNames) {
+      const event = this.events[eventName]
+
+      if (!event) {
+        console.log(`EmitNotifyEvents: Cannot find event '${eventName}'`)
+        continue
+      }
+
+      for (const templateName of event.customTemplates) {
+        const template = this.templates[templateName]
+
+        if (!template) {
+          console.log(`EmitNotifyEvents: Cannot find template '${templateName}'`)
+          continue
+        }
+
+        const customTemplates = await this.customTemplateModel.find({
+          where: {
+            templateName: {
+              equals: templateName
+            }
+          },
+          fields: ['sendOnEvents', 'id']
+        })
+
+        if (customTemplates.length === 0) {
+          console.log(`EmitNotifyEvents: Cannot find custom templates '${templateName}'`)
+          continue
+        }
+
+        const customTemplatesFilt = customTemplates.filter(t => Array.isArray(t.sendOnEvents) && t.sendOnEvents.includes(eventName))
+
+        if (customTemplatesFilt.length === 0) {
+          console.log(`EmitNotifyEvents: Cannot find custom templates '${templateName}' for event '${eventName}'`)
+          continue
+        }
+
+        for (const { id: customTemplateId } of customTemplatesFilt) {
+          const recipientsFromDatabase = await this.listCustomTemplateRecipients(customTemplateId)
+
+          const recipients = gatherRecipients(
+            recipientsFromStateMachine,
+            recipientsFromDatabase,
+            parameters
+          )
+
+          if (recipients.length === 0) {
+            console.log(`EmitNotifyEvents: Cannot find recipients for custom template '${customTemplateId}'`)
+            continue
+          }
+
+          const sendConfig = await this.sendConfigModel.find({
+            where: {
+              customTemplateId: {
+                equals: customTemplateId
+              }
+            }
+          })
+
+          if (sendConfig.length > 0 && sendConfig[0].offsetType) {
+            const stateMachineName = 'tymly_sendGovUkScheduledCustomTemplate_1_0'
+
+            let datetime
+
+            if (sendConfig[0].offsetType === 'HRS') {
+              console.log('offsetting by ', sendConfig[0].offsetLength, 'hours')
+              // datetime = moment().add(sendConfig[0].offsetLength, 'hours').toDate()
+              // Below line used during dev to make delay only 20 seconds
+              datetime = moment().add('20', 'seconds').toDate()
+            } else if (sendConfig[0].offsetType === 'DAYS') {
+              console.log('offsetting by ', sendConfig[0].offsetLength, 'days')
+              datetime = moment().add(sendConfig[0].offsetLength, 'days').toDate()
+            }
+
+            const config = {
+              stateMachineName,
+              input: {
+                recipients,
+                templateName,
+                customTemplateId
+              },
+              sendResponse: 'COMPLETE',
+              scheduleType: 'datetime',
+              datetime,
+              startOnBoot: false // todo
+            }
+
+            const taskResponse = await this.scheduleService.createScheduledTask(null, config)
+            await this.scheduleService.startTask(taskResponse.idProperties.key)
+          } else {
+            try {
+              const _sent = await this.templates[templateName].sendMessage(recipients, customTemplateId)
+              sent.push(_sent)
+            } catch (e) {
+              console.log('--------')
+              console.log('Failed to send message')
+              console.log(e)
+              console.log('--------')
+            }
+          }
+        }
+      }
+    }
+
+    return { sent }
+  }
 
   listMessageTemplates (options = {}) {
     const templates = []
@@ -150,9 +268,23 @@ class NotifyService {
   }
 }
 
+function gatherRecipients (fromStateMachine, fromDatabase, parameters) {
+  const recipients = []
+
+  for (const recipient of fromStateMachine) {
+    recipients.push({ recipient, parameters })
+  }
+
+  for (const { recipient } of fromDatabase) {
+    recipients.push({ recipient, parameters })
+  }
+
+  return recipients
+}
+
 module.exports = {
   schema,
   serviceClass: NotifyService,
   bootBefore: ['statebox'],
-  bootAfter: ['registry']
+  bootAfter: ['registry', 'schedule']
 }
